+++ libraries/extractor/src/main/java/androidx/media3/extractor/metadata/id3/Id3Decoder.java
@@ -60,6 +60,16 @@ public final class Id3Decoder extends SimpleMetadataDecoder {
   public static final FramePredicate NO_FRAMES_PREDICATE =
       (majorVersion, id0, id1, id2, id3) -> false;
 
+  /** A predicate that indicates artwork frames should not be decoded. */
+  public static final FramePredicate NO_ARTWORK_PREDICATE =
+      (majorVersion, id0, id1, id2, id3) -> {
+        if (majorVersion == 2) {
+          return !(id0 == 'P' && id1 == 'I' && id2 == 'C');
+        } else {
+          return !(id0 == 'A' && id1 == 'P' && id2 == 'I' && id3 == 'C');
+        }
+      };
+
   private static final String TAG = "Id3Decoder";
 
   /** The first three bytes of a well formed ID3 tag header. */