+++ libraries/extractor/src/main/java/androidx/media3/extractor/mp4/FragmentedMp4Extractor.java
@@ -114,7 +114,8 @@ public class FragmentedMp4Extractor implements Extractor {
         FLAG_READ_WITHIN_GOP_SAMPLE_DEPENDENCIES,
         FLAG_READ_WITHIN_GOP_SAMPLE_DEPENDENCIES_H265,
         FLAG_MERGE_FRAGMENTED_SIDX,
-        FLAG_READ_MFRA_FOR_SEEK_MAP
+        FLAG_READ_MFRA_FOR_SEEK_MAP,
+        FLAG_DISABLE_ARTWORK_METADATA
       })
   public @interface Flags {}
 
@@ -172,6 +173,9 @@ public class FragmentedMp4Extractor implements Extractor {
 
   /** Flag to enable reading the 'mfra' box for seeking in fragmented MP4s. */
   public static final int FLAG_READ_MFRA_FOR_SEEK_MAP = 1 << 9;
+
+  /** Flag to disable parsing of artwork metadata. */
+  public static final int FLAG_DISABLE_ARTWORK_METADATA = 1 << 10;
 
   /**
    * @deprecated Use {@link #newFactory(SubtitleParser.Factory)} instead.
@@ -817,7 +821,9 @@ public class FragmentedMp4Extractor implements Extractor {
     @Nullable Metadata udtaMetadata = null;
     @Nullable Mp4Box.LeafBox udta = moov.getLeafBoxOfType(Mp4Box.TYPE_udta);
     if (udta != null) {
-      udtaMetadata = BoxParser.parseUdta(udta);
+      udtaMetadata =
+          BoxParser.parseUdta(
+              udta, /* ignoreArtwork= */ (flags & FLAG_DISABLE_ARTWORK_METADATA) != 0);
       gaplessInfoHolder.setFromMetadata(udtaMetadata);
     }
     Metadata mvhdMetadata =